+++ docs/changelog.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Get links by contents, content substrings
 - Add opportunity to create sc-elements by SCs-text
 
+### Fixed
+- Improved tests work not only with the mock-server, but also with the real sc-machine server
+
 ## [v0.2.1]
 
 ### Fixed